Check the necessary tenants to be scraped based on the repo_map
This change filters for tenants that should be scraped based on the list
of repositories that are scraped.
So far, we were "scraping" (just update the tenant information in ES, no
"real" scraping) all tenants independently of the repositories that were
scraped. As we are now scraping the jobs per tenant, we should only
scrape the tenants which are relying on a repository that was changed -
as the repository may contain job definitions for this tenant.